+++ example/i2c_sht31.dart
@@ -14,6 +14,11 @@ void main() {
   // 1 for Raspberry Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)
   var i2c = I2C(1);
   try {
+    print("dart_periphery Version: $dartPeripheryVersion");
+    print("c-periphery Version   : ${getCperipheryVersion()}");
+    print('I2C info: ${i2c.getI2Cinfo()}');
+    print("SHT31 sensor");
+
     var sht31 = SHT31(i2c);
     print(sht31.getStatus());
     print('Serial number ${sht31.getSerialNumber()}');