+++ example/spi_bme280.dart
@@ -11,7 +11,11 @@ import 'package:dart_periphery/dart_periphery.dart';
 void main() {
   var spi = SPI(0, 0, SPImode.mode0, 1000000);
   try {
-    print('SPI info:${spi.getSPIinfo()}');
+    print("dart_periphery Version: $dartPeripheryVersion");
+    print("c-periphery Version   : ${getCperipheryVersion()}");
+    print('SPI info: ${spi.getSPIinfo()}');
+    print("BM280 sensor");
+
     var bme280 = BME280.spi(spi);
     var r = bme280.getValues();
     print('Temperature [°] ${r.temperature.toStringAsFixed(1)}');