+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.9.10 2024-11-??
+
+* Add SI1145 sunlight sensor
+* Rework README
+
 ## 0.9.9 2024-11-17
 
 * Add SDC30 - CO2, temperature amd humidity sensor